Assignments roles and checks

+++ app/actions/submission.ts
@@ -2,11 +2,30 @@
 
 // Import necessary modules
 import { PrismaClient } from '@prisma/client';
+import {getCurrentStudent} from "@/app/actions/getCurrentStudent";
 
 const prisma = new PrismaClient();
 
 
 export async function deleteSubmission(submissionId: string) {
+
+   const student = await getCurrentStudent();
+   if (!student) {
+       throw new Error('Not authenticated');
+    }
+    const submission = await prisma.assignmentSubmission.findUnique({
+        where: {
+            id: submissionId,
+        },
+    });
+
+    if (!submission) {
+        throw new Error('Submission not found');
+    }
+
+    if (submission.studentId !== student.id) {
+        throw new Error('You are not authorized to delete this submission');
+    }
     // Fetch the file details from the database
     const deleteFiles =  prisma.file.deleteMany({
         where: {
